refactor(scripts): split segments_ja.py into state machine + HTML helpers
Codex LOW L1 (file size 975 > 800 soft cap in CLAUDE.md coding style).

Extract HTML parsing helpers into a new `segments_ja_html.py` module:

- `split_table_cells` — pipe-table backslash-aware splitter
- `find_tag_end` — quote-aware HTML tag closer scanner
- `scan_for_matching_summary_close` — nesting-aware `</summary>` finder
- `tokenize_details_line` — `<details>`/`<summary>`/`</details>` line
  tokenizer for condensed one-liners
- `decode_html_entities` — 7-entity case-insensitive decode
- `html_inline_to_markdown_text` — inline HTML → markdown (preserves
  `<a href>` and `<code>` as invariant-token-bearing markdown syntax)
- `extract_html_table_cells` — `<tbody>` cell extraction with
  `<thead>` filtering

All four DETAILS/SUMMARY prefix regexes move to the helper module.
The main module keeps the `_DETAILS_TOKEN_RE` probe regex used by the
main-loop line dispatch (is this line worth tokenizing?) and the
line-dispatch regexes for fence/heading/callout/image/list/table/HR.

Line counts:
- segments_ja.py: 975 → 704 (under the 800-line soft cap)
- segments_ja_html.py: 356 (new, also under the cap)

No behavioural change — 390 tests still pass, mypy + ruff clean, and
288-page JA corpus still shows 141 byte-identical / 0 regressions.
